requirements.txt
django
django-admin-interface
django-crispy-forms
crispy-bootstrap4
pillow
zeep
psycopg2-binary
djangorestframework
requests
django-admin-confirm
django-axes
django-simple-captcha
django-recaptcha
django-multi-captcha-admin
cloudinary
whitenoise
xhtml2pdf
gunincore